+++ app/components/TeamSectionFull.tsx
