+++ app/components/TeamSectionHalf.tsx
